split error tests into their own module
